* fixed wrong return value for thread_wakeup * changed hwtimer_wait to use thread_sleep instead of mutexes

+++ core/include/tcb.h
@@ -34,7 +34,7 @@
 
 typedef struct tcb {
     char* sp;
-    unsigned int status;
+    uint8_t status;
 
     uint16_t pid;
     uint16_t priority;
